docs: Add hooks documentation (#94)
This pull request adds documentation for hooks in the `rustic`
repository. Hooks allow for the execution of custom scripts or commands
before or after certain commands, such as backup or restore. The
documentation provides an overview of the different types of hooks and
their configuration options. It also includes examples of how to use
hooks and run shell commands within hooks.

---------

Signed-off-by: simonsan <14062932+simonsan@users.noreply.github.com>
Co-authored-by: aawsome <37850842+aawsome@users.noreply.github.com>